Scripts state and listview
However bloc should be moved higher on the tree, because it will
be destroyed everytime the widget is destroyed. The unloaded state
is useless, because bloc is lazy, and will load only when needed.

There is notable code: factory constructor to initialize a list of
scripts as cubit state, and a static constructor async to initialize
scripts with their comment.